added support for TextureTransform in animation pointer + some refactoring of material animation pointer (WIP)
