+++ Runtime/Plugins/GLTFSerialization/Utilities/AnimationPointerUtilities.cs.meta
