+++ Runtime/Plugins/GLTFSerialization/Utilities/MaterialPropertyMapper.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: e35adb0d169a4046908867445f09f321
+timeCreated: 1708078060
